requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
boto3
unstructured
transformers
sentence-transformers
torch
python-dotenv